+++ api/src/bet/route/bet.routes.ts
@@ -60,6 +60,7 @@ export class BetRouter {
       ticket_number,
       page,
       limit,
+      group_id,
     } = req.query;
     const { user } = req;
     if (typeof date !== 'string') {
@@ -74,7 +75,13 @@ export class BetRouter {
     }
 
     try {
-      const organization_ids = await this.getOrgIds(req);
+      const allOrgIds = await this.getOrgIds(req);
+      const organization_ids =
+        typeof group_id === 'string' &&
+        req.user?.user.user_type !== USER_TYPE.CASHIER &&
+        allOrgIds.includes(group_id)
+          ? [group_id]
+          : allOrgIds;
 
       const result = await this.controller.getAllBets({
         date,
@@ -126,7 +133,7 @@ export class BetRouter {
   };
 
   private getTotalAmount: RequestHandler = async (req: Request, res: Response) => {
-    const { date, schedule_id, cashier_id, lottery_id } = req.query;
+    const { date, schedule_id, cashier_id, lottery_id, group_id } = req.query;
     const { user } = req;
     if (typeof date !== 'string') {
       const response: APIResponse<null> = {
@@ -140,7 +147,13 @@ export class BetRouter {
     }
 
     try {
-      const organization_ids = await this.getOrgIds(req);
+      const allOrgIds = await this.getOrgIds(req);
+      const organization_ids =
+        typeof group_id === 'string' &&
+        req.user?.user.user_type !== USER_TYPE.CASHIER &&
+        allOrgIds.includes(group_id)
+          ? [group_id]
+          : allOrgIds;
 
       const total = await this.controller.getTotalAmount({
         date,
@@ -185,7 +198,7 @@ export class BetRouter {
   };
 
   private getTotalPrize: RequestHandler = async (req: Request, res: Response) => {
-    const { date, schedule_id, cashier_id, lottery_id } = req.query;
+    const { date, schedule_id, cashier_id, lottery_id, group_id } = req.query;
     const { user } = req;
     if (typeof date !== 'string') {
       const response: APIResponse<null> = {
@@ -199,7 +212,13 @@ export class BetRouter {
     }
 
     try {
-      const organization_ids = await this.getOrgIds(req);
+      const allOrgIds = await this.getOrgIds(req);
+      const organization_ids =
+        typeof group_id === 'string' &&
+        req.user?.user.user_type !== USER_TYPE.CASHIER &&
+        allOrgIds.includes(group_id)
+          ? [group_id]
+          : allOrgIds;
 
       const total = await this.controller.getTotalPrize({
         date,
@@ -244,7 +263,7 @@ export class BetRouter {
   };
 
   private getAmountsByTicket: RequestHandler = async (req: Request, res: Response) => {
-    const { ticket_number } = req.query;
+    const { ticket_number, group_id } = req.query;
 
     if (typeof ticket_number !== 'string') {
       const response: APIResponse<null> = {
@@ -257,7 +276,13 @@ export class BetRouter {
       return;
     }
     try {
-      const organization_ids = await this.getOrgIds(req);
+      const allOrgIds = await this.getOrgIds(req);
+      const organization_ids =
+        typeof group_id === 'string' &&
+        req.user?.user.user_type !== USER_TYPE.CASHIER &&
+        allOrgIds.includes(group_id)
+          ? [group_id]
+          : allOrgIds;
 
       const total = await this.controller.getAmountsByTicket({
         ticket_number,