+++ web/src/components/settlement-payroll-table/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import Box from '@/components/box';
 import FilterSection from '@/components/filter-section';
 import CurrentAccountTable from '@/features/current-account/current-account-table';
@@ -8,11 +8,10 @@ import dayjs from 'dayjs';
 import toast from 'react-hot-toast';
 
 const SettlementPayrollTable = () => {
-  const [group, setGroup] = useState<string>('Todos');
   const [searchParams] = useSearchParams();
-  const { data, isLoading, isPending, isError, isSuccess } = useGetCurrentAccount(
-    searchParams.get('date')
-  );
+  const date = searchParams.get('date');
+  const group_id = searchParams.get('group_id');
+  const { data, isLoading, isPending, isError, isSuccess } = useGetCurrentAccount(date, group_id);
 
   useEffect(() => {
     if (isError) {
@@ -20,14 +19,11 @@ const SettlementPayrollTable = () => {
     } else if (isSuccess) {
       toast.success('Datos obtenidos correctamente');
     }
-  }, [isLoading, isPending]);
+  }, [isError, isSuccess]);
 
   return (
     <Box className="overflow-auto bg-[var(--primary-bg-content)] py-[24px] text-white">
-      <FilterSection
-        group={group}
-        onGroupChange={setGroup}
-      />
+      <FilterSection />
       <>
         Fecha de la Liquidación : {data?.[0]?.date && dayjs(data?.[0]?.date).format('DD/MM/YYYY')}
       </>